# test only: optimize MVCCValueMerger