--- a/go.mod
+++ b/go.mod
@@ -484,4 +484,4 @@ replace github.com/docker/docker => github.com/moby/moby v24.0.6+incompatible
 
 replace golang.org/x/time => github.com/cockroachdb/x-time v0.3.1-0.20230525123634-71747adb5d5c
 
-replace github.com/gogo/protobuf => github.com/cockroachdb/gogoproto v1.3.3-0.20241112020851-af1104e4bb72
+replace github.com/gogo/protobuf => github.com/cockroachdb/gogoproto v1.3.3-0.20241113180206-a3a930603783