#### go.mod: bump gogoproto dep

Epic: none
Release note: None

#### storage: optimize MVCCValueMerger

We use a sync.Pool to speed up allocation of `MVCCValueMerger` and to
reuse internal slices. We use updated gogoproto marshalling and
unmarshaling code for which avoids unnecessary allocations and allows
reuse of slices.

```
name               old time/op    new time/op    delta
MVCCValueMerger-4    16.6µs ± 1%     9.0µs ± 1%  -45.66%  (p=0.001 n=7+7)

name               old alloc/op   new alloc/op   delta
MVCCValueMerger-4    86.2kB ± 0%    27.9kB ± 0%  -67.61%  (p=0.000 n=8+8)

name               old allocs/op  new allocs/op  delta
MVCCValueMerger-4      37.0 ± 0%       3.0 ± 0%  -91.89%  (p=0.000 n=8+8)
```

Informs: #133313
Release note: None